Migrate world files
  - 20 image/texture stub files: replace INCLUDE_IMG/include_asset.h with
    assets/world.h OTR headers, inline Gfx[] display lists for dgb_01 imgs
  - 18 LOAD_ASSET files: replace ROM DMA loading with OTR in world.c,
    LoadPartyImage, lava_piranha, tst_13, osr_00, mgm_01/02, kmr_02/04/06/20/21/22,
    kzn_19, jan_04, iwa_01, dro_02, hos_05
  - 8 intptr_t files: 64-bit pointer-width casts in script_api (shops, rooms,
    push_blocks), ApplyTint, kkj_19, Credits, sam_05
  - world.c: OTR shape/texture loading via ResourceGetDataByName +
    Shape_LoadFromRawData, preserve DX improvements (ASSERT_MSG, DX_DEBUG_MENU,
    surface_set_walk_effect, phys_set_*, NODISCARD, heap allocators, SCREEN_*
    constants, get_map_IDs_by_name_checked)
  - hos_05/intro.c: OTR story page loading via ResourceGetDataByName,
    framebuffer sentinel fix, keep overrideSettings field name
  - Remove N64 BSS ordering hacks (MAP_STATIC_PAD, NOP_FIX, MAP_RODATA_PAD)

+++ src/world/script_api/push_blocks.c
@@ -409,7 +409,7 @@ API_CALLABLE(SetPushBlock) {
         s32 posY = blockGrid->centerPos.y;
         s32 posZ = blockGrid->centerPos.z + (gridZ * BLOCK_GRID_SIZE) + (BLOCK_GRID_SIZE / 2);
         blockEntityID = create_entity(&Entity_PushBlock, posX, posY, posZ, 0, 0, 0, 0, MAKE_ENTITY_END);
-        bind_trigger_1(&EVS_PushWall_PushBlock, TRIGGER_WALL_PUSH, blockEntityID + EVT_ENTITY_ID_BIT, (s32)blockGrid, blockEntityID, 3);
+        bind_trigger_1(&EVS_PushWall_PushBlock, TRIGGER_WALL_PUSH, blockEntityID + EVT_ENTITY_ID_BIT, (intptr_t)blockGrid, blockEntityID, 3);
         script->varTable[0] = blockEntityID;
     }
 
@@ -439,7 +439,7 @@ API_CALLABLE(FillPushBlockX) {
             s32 posY = blockGrid->centerPos.y;
             s32 posZ = blockGrid->centerPos.z + (gridZ * BLOCK_GRID_SIZE) + (BLOCK_GRID_SIZE / 2);
             blockEntityID = create_entity(&Entity_PushBlock, posX, posY, posZ, 0, 0, 0, 0, MAKE_ENTITY_END);
-            bind_trigger_1(&EVS_PushWall_PushBlock, TRIGGER_WALL_PUSH, blockEntityID + EVT_ENTITY_ID_BIT, (s32)blockGrid, blockEntityID, 3);
+            bind_trigger_1(&EVS_PushWall_PushBlock, TRIGGER_WALL_PUSH, blockEntityID + EVT_ENTITY_ID_BIT, (intptr_t)blockGrid, blockEntityID, 3);
             script->varTable[0] = blockEntityID;
         }
     }
@@ -471,7 +471,7 @@ API_CALLABLE(FillPushBlockZ) {
             s32 posY = blockGrid->centerPos.y;
             s32 posZ = blockGrid->centerPos.z + (gridZ * BLOCK_GRID_SIZE) + (BLOCK_GRID_SIZE / 2);
             blockEntityID = create_entity(&Entity_PushBlock, posX, posY, posZ, 0, 0, 0, 0, MAKE_ENTITY_END);
-            bind_trigger_1(&EVS_PushWall_PushBlock, TRIGGER_WALL_PUSH, blockEntityID + EVT_ENTITY_ID_BIT, (s32)blockGrid, blockEntityID, 3);
+            bind_trigger_1(&EVS_PushWall_PushBlock, TRIGGER_WALL_PUSH, blockEntityID + EVT_ENTITY_ID_BIT, (intptr_t)blockGrid, blockEntityID, 3);
             script->varTable[0] = blockEntityID;
         }
     }
